LPS-141305 Extract commerce logic to a new class
